requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
google-generativeai==0.3.1
pyaudio==0.2.13